requirements.txt
aws_cdk.aws_apigateway==1.148.0
aws_cdk.aws_lambda==1.148.0
aws_cdk.core==1.148.0